requirements.txt
asgiref==3.7.2
certifi==2024.2.2
Django==5.0.3
django-elasticsearch-dsl==8.0
django-elasticsearch-dsl-drf==0.22.5
django-nine==0.2.7
djangorestframework==3.14.0
drf-yasg==1.21.7
elastic-transport==8.12.0
elasticsearch==8.12.1
elasticsearch-dsl==8.12.0
inflection==0.5.1
packaging==23.2
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.10.0
uritemplate==4.1.1
urllib3==2.2.1
